Search Agent/requirements.txt
streamlit
langchain
duckduckgo-search
openai
python-dotenv
serpapi
requests
langchain-community
langchain-openai
google-search-results
